+++ ambari-web/app/mappers/alert_definitions_mapper.js
@@ -85,7 +85,7 @@ App.alertDefinitionsMapper = App.QuickDataMapper.create({
     interval: 'AlertDefinition.source.ams.interval'
   },
 
-  map: function (json) {
+  map: function (json, ignoreDelete) {
     console.time('App.alertDefinitionsMapper execution time');
     if (json && json.items) {
       var self = this,
@@ -125,19 +125,11 @@ App.alertDefinitionsMapper = App.QuickDataMapper.create({
 
         var convertedParameters = [];
         var sourceParameters = item.AlertDefinition.source.parameters;
-        if (Array.isArray(sourceParameters)) {
+        if (Ember.isArray(sourceParameters)) {
           sourceParameters.forEach(function (parameter) {
-            convertedParameters.push({
-              id: item.AlertDefinition.id + parameter.name,
-              name: parameter.name,
-              display_name: parameter.display_name,
-              units: parameter.units,
-              value: parameter.value,
-              description: parameter.description,
-              type: parameter.type,
-              threshold: parameter.threshold,
-              visibility: parameter.visibility
-            });
+            let hash = Em.getProperties(parameter, ['name', 'display_name', 'units', 'value', 'description', 'type', 'threshold', 'visibility']);
+            hash.id = item.AlertDefinition.id + parameter.name;
+            convertedParameters.push(hash);
           });
         }
 
@@ -231,21 +223,22 @@ App.alertDefinitionsMapper = App.QuickDataMapper.create({
         }
       }, this);
 
-      alertDefinitionsToDelete.forEach(function(definitionId) {
-        self.deleteRecord(existingAlertDefinitions.findProperty('id', definitionId));
-      });
+      if (!ignoreDelete) {
+        alertDefinitionsToDelete.forEach(function(definitionId) {
+          self.deleteRecord(existingAlertDefinitions.findProperty('id', definitionId));
+        });
+      }
 
       // load all mapped data to model
-      App.store.loadMany(this.get('reportModel'), alertReportDefinitions);
-      App.store.loadMany(this.get('parameterModel'), parameters);
-      App.store.loadMany(this.get('metricsSourceModel'), alertMetricsSourceDefinitions);
+      App.store.safeLoadMany(this.get('reportModel'), alertReportDefinitions);
+      App.store.safeLoadMany(this.get('parameterModel'), parameters);
+      App.store.safeLoadMany(this.get('metricsSourceModel'), alertMetricsSourceDefinitions);
       this.setMetricsSourcePropertyLists(this.get('metricsSourceModel'), alertMetricsSourceDefinitions);
-      App.store.loadMany(this.get('metricsUriModel'), alertMetricsUriDefinitions);
-      App.store.loadMany(this.get('metricsAmsModel'), alertMetricsAmsDefinitions);
-      // this loadMany takes too much time
-      App.store.loadMany(this.get('model'), alertDefinitions);
+      App.store.safeLoadMany(this.get('metricsUriModel'), alertMetricsUriDefinitions);
+      App.store.safeLoadMany(this.get('metricsAmsModel'), alertMetricsAmsDefinitions);
+      // this safeLoadMany takes too much time
+      App.store.safeLoadMany(this.get('model'), alertDefinitions);
       this.setAlertDefinitionsRawSourceData(rawSourceData);
-      App.store.commit();
     }
     console.timeEnd('App.alertDefinitionsMapper execution time');
   },
